+++ src/expandableCalendar/asCalendarConsumer.js
@@ -1,21 +1,13 @@
 import React, {Component} from 'react';
 import hoistNonReactStatic from 'hoist-non-react-statics';
-import CalendarContext from './calendarContext';
-
+import CalendarContext from './Context/calendarContext';
 
 function asCalendarConsumer(WrappedComponent) {
-  
   class CalendarConsumer extends Component {
     render() {
       return (
         <CalendarContext.Consumer>
-          {(context) => (
-            <WrappedComponent
-              ref={r => (this.contentRef = r)}
-              context={context}
-              {...this.props}
-            />
-          )}
+          {context => <WrappedComponent ref={r => (this.contentRef = r)} context={context} {...this.props} />}
         </CalendarContext.Consumer>
       );
     }
