--- src/expandableCalendar/calendarContext.js
@@ -1,4 +1,0 @@
-import React from 'react';
-
-const CalendarContext = React.createContext({});
-export default CalendarContext;
