+++ src/index.js
@@ -4,7 +4,7 @@ export {default as Agenda} from './agenda';
 export {default as ExpandableCalendar} from './expandableCalendar';
 export {default as AgendaList} from './expandableCalendar/agendaList';
 export {default as Timeline} from './timeline/Timeline';
-export {default as CalendarProvider} from './expandableCalendar/calendarProvider';
+export {default as CalendarProvider} from './expandableCalendar/Context/calendarProvider';
 export {default as asCalendarConsumer} from './expandableCalendar/asCalendarConsumer';
 export {default as WeekCalendar} from './expandableCalendar/WeekCalendar';
 export {default as LocaleConfig} from 'xdate';
